change component position
