+++ src/s2-features/f4-passwordRecovery/pr3-dal/passwordRecoveryAPI.ts
@@ -1,9 +1,7 @@
 import { instance } from '../../../s1-main/m3-dal/instance'
 
 const messageView = (token: string) => {
-   return `<div style="background-color: lime; padding: 15px">
-            <a href=http://localhost:3000/#/new-password-entry/${token}'>
-            link</a></div>`
+   return `<div style="background-color: lime; padding: 15px"><a href=http://localhost:3000/#/new-password-entry/${token}'>Ссылка на восстановление пароля</a></div>`
 }
 
 export const passRecoveryAPI = {
